Moved Paintbrush.js to js/Puppet
